requirements.txt
aiogram==3.13.1
yt-dlp==2024.10.22
requests==2.32.3
aiohttp==3.10.10
aiosignal==1.3.1
frozenlist==1.5.0
attrs==24.2.0
async-timeout==4.0.3
python-dotenv==1.0.1
sqlite-utils==3.37



